Cleaned up and added another test texture.
