+++ src/CMakeLists.txt
@@ -12,13 +12,12 @@ source_group("Shaders" FILES ${SHADERS})
 
 # Build and link executable.
 add_executable ( ${PROJECT_NAME}
-    main.cpp
-    ${SHADERS}
-    RenderView.h
-    RenderView.cpp
-    temp_screen_helper.cpp
-    temp_screen_helper.h
-    Application.h)
+        main.cpp
+        ${SHADERS}
+        RenderView.h
+        RenderView.cpp
+        Application.h
+        Application.cpp)
 
 target_link_libraries ( ${PROJECT_NAME} labhelper )
 config_build_output()
